Move all IBAN related code inside iban.go
